Implemented the onCreateViewHolder method in SleepNightAdapter and added a setter field for the data that notifies the RecyclerView that the data set has changed.
Created text_item_view to give to our recycler view

Made the nights variable public in the SleepTrackerVewModel

Added adapter to SleepTrackerFragment
Bound the adapter to the sleeplist recycler view in the fragment_sleep_tracker layout
Added observer to nights variable in the view model which will update our adapter data variable in the SleepTrackerAdapter, which in turn notify the RecyclerView that the data set has changed and it needs to redraw.

Added a few string resources.